requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
boto3==1.26.136
botocore==1.29.136
celery==5.2.7
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==4.2.1
django-cors-headers==4.0.0
django-storages==1.13.2
djangorestframework==3.14.0
jmespath==1.0.1
kombu==5.2.4
mysql-connector-python==8.0.33
prompt-toolkit==3.0.38
protobuf==3.20.3
python-dateutil==2.8.2
pytz==2023.3
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
vine==5.0.0
wcwidth==0.2.6
gunicorn
